Convert ApiGenerator from an object to a class
This change allows the `ApiGenerator` to add properties and so avoid
the churn of having to pass additional parameters down to where they
are needed. It is needed for a follow up change that will make
`SignatureFileLoader` a class and so require an instance to be passed
to where it is used.

Bug: 301405075
Test: ./gradlew
Change-Id: I119a5b2c367aefe2f997e16693204629b6f7052c