FlaggedApi: do not require flags on inherited fields
Fixes an issue where flags were erroneously
required on duplicated inherited fields.

This lint can currently only erroneously happen when
we're adding a new class that inherits a field from
a super class or interface, where that class is
referencable but not emittable.

Note that this does not currently happen with the
public API surface, because in that case we only
duplicate the method for hidden super classes, and in
that case the lint is justified.

Bug: 297362755
Test: ./gradlew :metalava:test --tests ApiLintTest.*Flagged*
Change-Id: Ia409386770747368d00af3ebdac2297526a1cd15