Correct possible cache corruption in SignatureFileCache
Previously, the `load()` method took three parameters that could affect
the contents of the returned `TextCodebase` but only used the `File` in
the key. That meant it was possible to load the same file twice with
different values for the other parameters, which should return
different `TextCodebase` instances, but would actually return the same
one; i.e. whichever was loaded first.

This change fixes that by:
1. Taking the `annotationManager` as a constructor parameter; this was
   possible because all callers used the same one.
2. Adding the `classResolver` to the cache key; this was necessary as
   while most calls just used the default of `null`, at least one call
   passes in its own.

Adding a constructor parameter would have affected the default value in
`ActionContext` but rather than pass `annotationManager` to
`ActionContext` (which none of the extension functions need) this just
removed the `signatureFileCache` from it as only those extension
functions which are private to the `Driver.kt` file need it. Instead,
the `signatureFileCache` was added as a parameter to those functions.

Bug: 301405075
Test: ./gradlew
Change-Id: Ib500bf415c0b2785edf8473d627f632cc6df223f