Shift some ifs in CompatibilityCheck
Changes this from one if with four clauses to nested ifs.

The following CL will change the if conditions from null checks to
using `is VariableTypeItem`. Using the nested ifs means kotlin
can detect what the else condition is, and there won't need to be
any casting to VariableTypeItem after the if checks.

Bug: 301285165
Test: existing tests pass
Change-Id: I8fc4190964db47f747423bf4fe61bb8760cf4c88