Simplify handling of baseline in DefaultReporter
Previously, when `customBaseline` was `null` it would use the
`options.baseline` because that was only set during `Options.parse`
which was some time after the `Options.reporter` was created. However,
previous changes fixed that problem by creating a separate
`bootstrapReporter` that is used before `Options.parse` and moved the
creation of `Options.reporter` into `Option.parse`. So, the `baseline`
object is now available at the time that `Options.reporter` is created
so can just be passed into the constructor as a normal property,
removing the use of `options.baseline`.

This change replaces `customBaseline` and `getBaseline()` with a simple
`val baseline` constructor property. That allows the
`Options.baseline*` properties to be replaced with local variables in
`Options.parse`.

Bug: 295013652
Test: ./gradlew
Change-Id: I5401c58e81517257dc821aff3bd387dc0f7ce821