Stop DefaultReporter using options.stubPackages
Previously, the `options.stubPackages was used because it was
initialized after `Options.reporter` was created but that is no longer
the case. While the IssueReportingOptions.bootstrapReporter` is created
before `options.stubPackages` is initialized it does not need to use
the `PackageFilter` as that only affects issues reported on model
`Item`s and they are created after the `bootstrapReporter` is no longer
used to report issues.

So, this change adds a `packageFilter` constructor property to
`DefaultReporter` and passes in `stubPackages` to the non-bootstrap
reporters. That allowed `stubPackages` to be marked private.

Bug: 295013652
Test: ./gradlew
Change-Id: Ie854d4e3a774e611093e685ae778f3cfe64aee37