Annotate metalava APIs used from outside metalava
In order to prevent APIs being removed that are used from outside
Metalava this change annotates those API elements with the
`@MetalavaApi` annotation. An API element annotated with that must
not be removed or changed in such a way as to break Java source
compatibility, e.g. while adding a new parameter with a default value
does not break Kotlin source compatibility it does break Java source
compatibility.

The process for determining which APIs needed annotation was simple.
Manually review the `ApprovedApis.java` file and annotate every class,
method and field that is used by that code.

Bug: 302654813
Test: ./gradlew
Change-Id: I65a260603a565e3fc0ca8d3ec8a6749c37eb6c23