Rewrite text array type parsing to prepare for modifiers
Previously, when parsing an array this recurred on parsing the
component type for each level of a multi-dimensional array.
However, because of how type-use annotations appear on arrays, the
simple recursive approach won't correctly handle matching which level
of the array should have which annotations.

To address the annotation issue, when parsing a multi-dimensional array
this first gathers the annotations/nullability for each array type,
parses the inner component type, and then builds up the array types.

Bug: 300081840
Test: Added test which inputs and outputs a signature file to verify
that reconstructing the array type string works as expected.

Change-Id: I2bb60ae3c0003816382d5ebd215e41fe468bc85b