Assume type variables without explicit bounds extend Object
Previously, `TextTypeItem.toErasedTypeString()` would leave references
to a type variable `<T>` that did not have an explicit bound as `T`
instead of replacing it with `java.lang.Object` as `PsiTypeItem` does.
There was some code to support that already in metalava but it was
hidden inside a conditional that was always `false` and so it was never
used.

This change inlines that previously ignored code which fixes some of
the suite tests. It was also necessary to modify some other tests
which while not testing `findMethod(String,String)` used it as part of
the test and relied on the previous behavior.

Bug: 303839034
Test: ./gradlew
Change-Id: I019fa5e8aba6bb9f09fc9c16192f1eacee188716