Simplify handling of superConstructor in JavaStubWriter
A long time ago, the constructors that were present in the sources were
written out into the stubs which meant that the stubs writer needed to
handle delegation to both super constructors and other constructors
within the same class, so it could write out `this(...)` or
`super(...)` as required. However, that caused problems with internal
classes leaking out into the APIs so that was simplified to always
delegate to the super class's default `stubConstructor`.

This change cleans up the now unused logic in `JavaStubWriter` for
calling `this(...)` and updates the documentation for
`superConstructor`.

Bug: 304226685
Test: ./gradlew
Change-Id: I5500217a34b95f08fb163c0703f60e55334b6e33