Remove adding of concrete overrides of abstract methods
That functionality does not work when
`apiClassResolution == ApiClassResolution.API_CLASSPATH` as it attempts
to cast a `PsiClassItem` to ` TextClassItem`. However, that should no
longer be needed as concrete overrides of abstract methods are now kept
in the signature files (or at least should be soon). So, rather than
try and get that functionality working with a mixture of text and psi
model classes this just removes it.

Bug: 303687754
Test: ./gradlew
Change-Id: I086939123277e1fe0c0e229208d18e45eff58ee5