Remove unused `strict` parameter from Item.containingClass(...)
The `strict` parameter would only have an effect when the method was
called on a `ClassItem` and `strict` was set to `false`. However,
`strict` defaulted to `true` and the only place where it was `false`
was when it was called from `ParameterItem` on a `MethodItem` which
meant it was ignored. As, the `strict` parameter had no effect this
change removed it.

Bug: 302654813
Test: ./gradlew
Change-Id: I9f8810d254b757215e26f45744f61421fc9f0d45