Remove `strict` parameter from Item.containingPackage(...)
The `strict` parameter would only have an effect when the method was
called on a `PackageItem` and `strict` was set to `false`. However,
`strict` defaulted to `true` and in all but one of the places where it
was `false` the method was not being called on something which could
have been `PackageItem` so it would have had no effect. The exception
was in the `DefaultReporter` where it was used to get the `PackageItem`
for filtering purposes.

This change removes the `strict` parameter and modified th
`DefaultReporter` to handle `PackageItem` specially.

Bug: 302654813
Test: ./gradlew
Change-Id: I02fd8f370e176851fbdc52f2a8fdd2571577dcb0