Remove unused extension function Item.containingClass()
This extension function was shadowed by the
`containingClass(strict: Boolean = true)` method that was added to
`Item` directly.

Usually, the compiler/IDE would report a warning when an extension
function will not be called because it is shadowed by an actual method
on the type being extended but it seems that in this case it did not
because of the `strict` parameter. It seems that even though that
parameter has a default and so the method has the same effective
signature as the extension function the compiler/IDE treat them as
having different signatures.

Before removing I triple checked that the extension function was not
being used:
1. I modified the extension function to throw an exception but none of
   the tests failed. However, it could have been the case that the
   tests did not cover the calls to the extension function.
2. To check the test coverage I temporarily removed the default value
   for the `strict` parameter and the IDE indicated that the extension
   function was now being used in two places. I renamed the extension
   function so it was no longer shadowed and modified those two places
   to use the extension function (which still threw an exception) and
   reran the tests. A lot of them failed due to that exception being
   thrown.
3. Finally, I removed the `strict` parameter altogether and the IDE
   indicated that the extension function was being shadowed.

Bug: 302654813
Test: ./gradlew
Change-Id: Ie1769edb7dafc7dd60d38ef7340a91afa1ca37a2