Convert SignatureFileLoader from an object to a class
Previously, it was just an object which meant that all the parameters
had to be passed to each call. This change switches it to a class to
allow some to be provided as properties. This will make it easier to
add new properties.

The `ClassResolver` is kept as a property because that can differ by
file whereas `AnnotationManager` does not.

Bug: 305396465
Test: ./gradlew
Change-Id: I3a050ea4b8513886737536a1db2bc02bc9157387