Allow parsing of legacy signature files
Previously, when parsing a legacy signature file, i.e. a signature file
that uses a format that is no longer supported, e.g. `FileFormat.V1`
the `FileFormat.parseHeader(...)` function would always fail. This adds
a `formatForLegacyFiles`, which allows callers to force the parser to
try and parse the legcay files as the supplied format.

This is provided primarily to allow the parsing of legacy API files
that provided no header, i.e. `FileFormat.V1` by treating tham as if
they were `FileFormat.V2` which is essentially what metalava did before
it removed support for `FileFormat.V1`.

Bug: 305396465
Test: ./gradlew
Change-Id: I107c570cade00f17c50eadf7ea9530c9f6512c24