Don't throw away unknown platform annotations
Previously metalava would ignore any annotations in source code that
were not explicitly listed to be kept. Change this default to not ignore
them.

The purpose of this is to avoid having to modify metalava every time
an annotation is added which has e.g. @memberDoc in it. As far as I can
tell, there's no obvious downside to keeping track of all annotations
in metalava's internal representation. There is a separate bit of config
that decides where the annotation is output (stubs, external, docs etc),
and the current default of "do not output anywhere" remains.

Bug: 305930201
Test: m ds-docs-java before and after & observe diffs (memberDoc from a
      few different annotations appear, whereas they did not before)
Change-Id: I8617136cacd941f3d07cd4d7884e63de8d1f4b21