Consider type parameter bounds when comparing method overrides
Previously, a type parameter's bounds were not considered when
comparing times in `MethodItem.sameSignature`. That meant that methods
that used a type parameter with the same name as a super method would
be considered to have the same signature as that super method, even if
the type parameters had different bounds.

In the following `EnumMap.put(K,V)` would have been considered to have
the same signature as `AbstractMap.put(K,V)`.
```
public abstract class AbstractMap<K,V> implements Map<K,V> {
    public V put(K key, V value) { ... }
}
public class EnumMap<K extends Enum<K>, V> extends AbstractMap<K, V> {
    public V put(K key, V value) { ... }
}
```

Yet when the generics are erased it is clear that they are not the
same as one is `EnumMap.put(Enum,Object)` and the other is
`AbstractMap.put(Object,Object)`:
```
public abstract class AbstractMap implements Map {
    public Object put(Object key, Object value) { ... }
}
public class EnumMap extends AbstractMap {
    public Object put(Enum key, Object value) { ... }
}
```

This change modifies `sameType` to only consider two types as being
the same if both their unerased and erased types are the same. This is
not particularly clean, or efficient but until the `TypeItem` hierarchy
has been fully implemented it is the simplest way to fix the behavior.

It purposely does not just check the erased types because that would
cause significant churn and also cause the removal of overrides that
add annotations, like nullability annotations.

This change only takes effect when `add-additional-overrides=yes`.

Bug: 306333334
Test: ./gradlew
Change-Id: I8dac63dcba39d093076eaf5897b85f7599373642