Modify Turbine Initializer to incorporate Binder
Earlier we were using Turbine's binder only for setting up Class
Hierarcy and others parts were completed using parsed units. However
it will not feasible to continue with this further as implementation of
Annotations and Types will require symbols to be resolved.
This change aims to use Binder's output for initialization and now
creates the items and sets up hierarchy in single pass instead of two

Bug: 307225873
Test: ./gradlew
Change-Id: I4cafe770b664ac2c62dfd686d019687ecce53a7d