Remove usages of RegistrationName lint
The API guidelines were updated (go/android-api-guidelines#registering-unregistering-callbacks) and this rule no longer applies, register/unregister and add/remove are both allowed for both Listener and Callback methods. This now just applies the same PairedRegistration check in both cases.

This leaves the issue defined in Issues.kt because it is referenced in baselines for both AndroidX and platform, so removing it entirely would cause an error when parsing the baselines.

Bug: 303638394
Test: Updated existing tests to now check PairedRegistration for add/Callback and register/Listener
Change-Id: I0a3491c97f87a46943de98e51fe10556ea5f7413