Remove --dex-api
Removes support for `--dex-api` from the main metalava command and
moves the `DexApiWriter` into the `...metalava.cli.signature` package
alongside the `SignatureToDexCommand` that is the only use of it. A
couple of uses of `--dex-api` were also removed but that did not reduce
the test coverage as that was already duplicated in
`SignatureToDexCommandTest`.

Bug: 307737035
Test: ./gradlew
Change-Id: Ia81c9f64040e89219f0e73873c1e455e1c2451f6