Add tests for methods with polymorphic signatures
The java compiler treats methods with polymorphic signatures specially.
It identifies a method as being polymorphic according to the rules
defined in JLS 15.12.3.

See https://docs.oracle.com/javase/specs/jls/se21/html/jls-15.html#jls-15.12.3
for the latest (at time of writing rules). They state:

> A method is signature polymorphic if all of the following are true:
>
> * It is declared in the java.lang.invoke.MethodHandle class or the
>   java.lang.invoke.VarHandle class.
>
> * It has a single variable arity parameter (§8.4.1) whose declared
>   type is Object[].
>
> * It is native.

The latter point means that the `native` modifier is an important part
of a polymorphic method's signature which is a problem as Metalava
currently excludes the `native` modifier from the methods which means
that the compiler does not handle them properly resulting in runtime
errors. This change adds some tests to cover the handling of those
methods and show the existing, and broken behavior. A follow up change
will fix the methods.

Bug: 308022433
Test: ./gradlew
Change-Id: I03b879e500c1bdf30e7c9603cd52195bb4ded780