Make apiSince metadata use the same filter as signature files
Previously, the default ApiVisitor filters were being used, which did not filter the same elements as the signature filters. This was an issue because to reduce the number of metalava invocations, we're passing all previous APIs as signature files but the current API as the source code, so APIs that don't appear in signature files (like overriding methods) were showing up in the apiSince metadata as added in the latest version.

ApiVisitor has two constructors, one which requires `filterEmit` and `filterReference`, and one which does not have them as parameters and constructs values for them based on the other constructor inputs. There are two use cases for `addApisFromCodebase`, apiSince and the api levels xml, which would have needed to use different constructors, so one of the constructors has been updated to make `filterEmit` and `filterReference` optional.

Bug: 298405576
Test: Added test that an overriding method which does not appear in signature files also does not appear in the generated json anymore, and one making sure experimental APIs still appear
Change-Id: I8321679c68340ff71edc79e7a070bd37747ffe9b